Estava ocorrendo erro ao inicializar o BD. Removido Contexto da API. MVC e API utilizam o mesmo contexto para o Identity. Removido o Identity do contexto do ClickMarket. Identity utiliza o AppDbContext exclusicamente. Removida as duplicidades de código ao inicializar o BD, concentrado na camada de dados. Recriada as migrations, uma para cada contexto. Program reorganizada em classes específicas.
