+++ src/BackEnd/ClickMaket.Data/Configurations/DatabaseSelect.cs
@@ -1,0 +1,35 @@
+﻿using ClickMarket.Data.Context;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+namespace ClickMarket.Data.Configurations
+{
+    public static class DatabaseSelect
+    {
+        public static WebApplicationBuilder AddDatabaseSelector(this WebApplicationBuilder builder)
+        {
+            if (builder.Environment.IsDevelopment())
+            {
+                var connectionString = builder.Configuration.GetConnectionString("DefaultConnectionLite") ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
+                builder.Services.AddDbContext<ApplicationDbContext>(options =>
+                options.UseSqlite(connectionString));
+
+                builder.Services.AddDbContext<ClickDbContext>(options =>
+                    options.UseSqlite(connectionString));
+            }
+            else
+            {
+                var connectionString = builder.Configuration.GetConnectionString("DefaultConnection") ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
+                builder.Services.AddDbContext<ApplicationDbContext>(options =>
+                    options.UseSqlServer(connectionString));
+
+                builder.Services.AddDbContext<ClickDbContext>(options =>
+                    options.UseSqlServer(connectionString));
+            }
+            return builder;
+        }
+    }
+}
