+++ src/FrontEnd/ClickMarket.Api/Configurations/ApiConfig.cs
@@ -1,4 +1,6 @@
-﻿namespace ClickMarket.Api.Configurations;
+﻿using ClickMarket.Data.Context;
+
+namespace ClickMarket.Api.Configurations;
 
 public static class ApiConfig
 {
