+++ src/FrontEnd/ClickMarket.Api/Configurations/DatabaseSelectExtension.cs
@@ -1,4 +1,5 @@
 ﻿using ClickMarket.Api.Context;
+using ClickMarket.Data.Configurations;
 using ClickMarket.Data.Context;
 using Microsoft.EntityFrameworkCore;
 
@@ -8,26 +9,7 @@ namespace ClickMarket.Api.Configurations
     {
         public static WebApplicationBuilder AddDatabaseSelector(this WebApplicationBuilder builder)
         {
-            if (builder.Environment.IsDevelopment())
-            {
-                var connectionString = builder.Configuration.GetConnectionString("DefaultConnectionLite") ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
-                builder.Services.AddDbContext<ApiDbContext>(options =>
-                options.UseSqlite(connectionString));
-
-                builder.Services.AddDbContext<ClickDbContext>(options =>
-                    options.UseSqlite(connectionString));
-            }
-            else
-            {
-                var connectionString = builder.Configuration.GetConnectionString("DefaultConnection") ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
-                builder.Services.AddDbContext<ApiDbContext>(options =>
-                    options.UseSqlServer(connectionString));
-
-                builder.Services.AddDbContext<ClickDbContext>(options =>
-                    options.UseSqlServer(connectionString));
-            }
-
-            return builder;
+            return DatabaseSelect.AddDatabaseSelector(builder);
         }
     }
 }
