+++ src/FrontEnd/ClickMarket.AppMvc/Configurations/IdentityConfig.cs
@@ -1,0 +1,17 @@
+﻿using ClickMarket.Data.Context;
+using Microsoft.AspNetCore.Identity;
+
+namespace ClickMarket.AppMvc.Configurations
+{
+    public static class IdentityConfig
+    {
+        public static WebApplicationBuilder AddIdentityConfig(this WebApplicationBuilder builder)
+        {
+            builder.Services.AddDefaultIdentity<IdentityUser>(options => options.SignIn.RequireConfirmedAccount = false)
+                .AddEntityFrameworkStores<ApplicationDbContext>()
+                .AddDefaultTokenProviders();
+
+            return builder;
+        }
+    }
+}
