+++ src/FrontEnd/ClickMarket.AppMvc/Data/ApplicationDbContext.cs
@@ -3,11 +3,11 @@ using Microsoft.EntityFrameworkCore;
 
 namespace ClickMarket.AppMvc.Data
 {
-    public class ApplicationDbContext : IdentityDbContext
-    {
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
-            : base(options)
-        {
-        }
-    }
+    //public class ApplicationDbContext : IdentityDbContext
+    //{
+    //    public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
+    //        : base(options)
+    //    {
+    //    }
+    //}
 }
